renamed some stuff. still working
