+++ src/main/java/components/FontRenderer.java
@@ -6,7 +6,7 @@ public class FontRenderer extends Component {
 
     @Override
     public void start() {
-        if (gameObject.getComponent(SpriteRenderer.class) != null) {
+        if (gameObject.getComponent(TriangleRenderer.class) != null) {
             System.out.println("Found Font Renderer!");
         }
     }
